requirements.txt
google-generativeai==0.8.1
google-ai-generativelanguage==0.6.9
gradio==4.44.0